C05 fixed

+++ C05/ex06/ft_is_prime.c
@@ -6,7 +6,7 @@
 /*   By: cari <cari@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/11 17:39:38 by cari              #+#    #+#             */
-/*   Updated: 2024/02/11 21:06:44 by cari             ###   ########.fr       */
+/*   Updated: 2024/02/13 21:03:18 by cari             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int	ft_is_prime(int nb)
 	div = 2;
 	if (nb <= 1)
 		return (0);
-	while (div * div <= nb)
+	while ((long) div * div <= (long) nb)
 	{
 		if (nb % div == 0)
 			return (0);
